Simplify condition

+++ custom_news_feed/custom_news_feed.py
@@ -551,15 +551,14 @@ class CustomNewsFeed:
     def remove_deprecated_hashfiles(self):
         """Removes hash files that are not in the current news anymore"""
         for file_path in os.listdir(self.settingspath):
-            if os.path.isfile(os.path.join(self.settingspath, file_path)):
-                if file_path != self.previous_news_name:
-                    stillExists = False
-                    for article in self.news["NewsArticles"]:
-                        if article["Hash"] == file_path:
-                            stillExists = True
-                            break
-                    if stillExists == False:
-                        self.delete_hashfile(file_path)
+            if file_path != self.previous_news_name and os.path.isfile(os.path.join(self.settingspath, file_path)):
+                stillExists = False
+                for article in self.news["NewsArticles"]:
+                    if article["Hash"] == file_path:
+                        stillExists = True
+                        break
+                if stillExists == False:
+                    self.delete_hashfile(file_path)
 
     def create_tab_widget(self, tab):
         """ Creates a layout for the news articles and adds it to the tab """
